resolves review feedback
