+++ src/frontend/src/lib/components/tokens/ManageTokenToggle.svelte
@@ -1,8 +1,9 @@
 <script lang="ts">
 	import { Toggle } from '@dfinity/gix-components';
 	import { createEventDispatcher } from 'svelte';
-	import { type EthereumUserToken, isEthereumUserTokenDisabled } from '$eth/types/erc20-user-token';
+	import { type EthereumUserToken } from '$eth/types/erc20-user-token';
 	import { i18n } from '$lib/stores/i18n.store';
+	import { isEthereumUserTokenDisabled } from '$lib/utils/token-toggle.utils';
 
 	export let token: EthereumUserToken;
